Final updates: comments, logs and tests

+++ src/main/java/com/example/blaze/test/service/ProductServiceImpl.java
@@ -2,6 +2,7 @@ package com.example.blaze.test.service;
 
 import com.example.blaze.test.domain.Product;
 import com.example.blaze.test.dto.ProductDto;
+import com.example.blaze.test.repository.OrderRepository;
 import com.example.blaze.test.repository.ProductRepository;
 import org.springframework.beans.BeanUtils;
 import org.springframework.data.domain.Page;
@@ -13,17 +14,19 @@ import org.springframework.stereotype.Service;
 public class ProductServiceImpl implements ProductService {
 
     private final ProductRepository productRepository;
+    private final OrderRepository orderRepository;
     private final SequenceGeneratorService sequenceGeneratorService;
 
-    public ProductServiceImpl(ProductRepository productRepository, SequenceGeneratorService sequenceGeneratorService) {
+    public ProductServiceImpl(ProductRepository productRepository, OrderRepository orderRepository, SequenceGeneratorService sequenceGeneratorService) {
         this.productRepository = productRepository;
+        this.orderRepository = orderRepository;
         this.sequenceGeneratorService = sequenceGeneratorService;
     }
 
     @Override
     public Page<ProductDto> getAll(Integer page, Integer size) {
         Pageable pageable = PageRequest.of(page, size);
-        return productRepository.findAll(pageable).map(this::toProductDto);
+        return productRepository.findPagedProducts(pageable).map(this::toProductDto);
     }
 
     @Override
@@ -34,6 +37,7 @@ public class ProductServiceImpl implements ProductService {
     @Override
     public ProductDto create(ProductDto productDto) {
         var product = this.toProduct(productDto);
+        product.setId(null);
         product.setId(String.valueOf(sequenceGeneratorService.generateSequenceProduct(Product.SEQUENCE_NAME)));
         return this.toProductDto(productRepository.save(product));
     }
@@ -50,11 +54,23 @@ public class ProductServiceImpl implements ProductService {
                 .orElse(null);
     }
 
+    /*
+        Deletes product with id=:id
+        Verifies if product is being used by some order
+     */
     @Override
     public void delete(String id) {
-        this.productRepository.deleteById(id);
+        if (this.orderRepository.findOrderWithProduct(id).isEmpty()) {
+            this.productRepository.deleteById(id);
+        } else {
+            throw new IllegalArgumentException("Product with id=" + id + " is currently being used by orders");
+        }
     }
 
+    /*
+        MAPPERS:
+        For Dto to domain and the other way round
+     */
     private Product toProduct(ProductDto dto) {
         var p = new Product();
         BeanUtils.copyProperties(dto, p);
